fix(web): extract computeQualityChecks from 'use client' file to fix RSC boundary error
Skill detail page 500'd because the pure function lived in a 'use client'
module — Turbopack proxied it as a client reference, which explodes on SSR.
